backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["plein <philipp@entwicklerheld.de>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.1"
dj-database-url = "^2.1.0"
sentry-sdk = "^1.29.2"


websockets = "^11.0.3"
redis = "^5.0.0"
async-timeout = "^4.0.3"
aiopg = "^1.4.0"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.25.0"
fastapi = "^0.109.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
docker = "^6.0.1"
httpx = "^0.26.0"
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
